+++ src/components/MainPage/store/main-page.api.js
@@ -1,8 +1,8 @@
 import axios from "axios";
 
-const getUsersList = async (url) => {
+const getUsersList = async (pageNumber) => {
     return await axios
-        .get(url)
+        .get(`http://localhost:5000/api/users/?page=${pageNumber}`)
         .then((response) => response.data)
         .catch((error) => console.error(error));
 };
