+++ src/components/MainPage/store/main-page.saga.js
@@ -1,9 +1,6 @@
-import { call, put, takeEvery } from "redux-saga/effects";
-import {
-    GET_USERS_REQUESTED,
-    getUsers,
-    setTotalPages,
-} from "./main-page.action";
+import { call, put, takeLatest } from "redux-saga/effects";
+import { GET_USERS_REQUESTED } from "./main-page.action";
+import { getUsers,setTotalPages } from "./main-page.action-creators";
 import adapter from "../../../common/utils";
 import getUsersList from "./main-page.api";
 
@@ -14,7 +11,7 @@ function* getUserData({ payload }) {
 }
 
 function* mainPageSaga() {
-    yield takeEvery(GET_USERS_REQUESTED, getUserData);
+    yield takeLatest(GET_USERS_REQUESTED, getUserData);
 }
 
 export default mainPageSaga;
